feat: vault mint routes for OETH and OUSD
